Atualização Unidades (Funcionando)

+++ NexaLog-Backend/NexaLog-Backend/Models/Lote.cs
@@ -1,9 +1,8 @@
 using NexaLog_Backend.Models;
-
 public class Lote
 {
     public int IdLote { get; set; }
-    public int QuantidadeLote { get; set; }
+    public decimal QuantidadeLote { get; set; }
     public string CodLote { get; set; } = string.Empty;
     public DateOnly DataValidade { get; set; }
     public DateOnly DataFabricacao { get; set; }
